add Loans component
