+++ src/components/App/App.tsx
@@ -3,7 +3,7 @@ import { CssBaseline, ThemeProvider } from '@mui/material';
 import theme from '../../theme/theme';
 import Header from '../Header/Header';
 import Main from '../Main/Main';
-import LoansList from '../Loans/LoansList';
+import Loans from '../Loans/Loans';
 
 function App() {
   return (
@@ -12,7 +12,7 @@ function App() {
       <ThemeProvider theme={theme}>
         <Header />
         <Main>
-          <LoansList />
+          <Loans />
         </Main>
       </ThemeProvider>
     </>
